pyproject.toml
[project]
name = "infera"
version = "0.1.0"
description = "Agentic infrastructure provisioning from code analysis"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Pankaj Garkoti" }
]
keywords = ["infrastructure", "cloud", "gcp", "aws", "azure", "terraform", "ai", "agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
]

dependencies = [
    # CLI
    "typer>=0.9.0",
    "rich>=13.0.0",

    # Claude Agent SDK
    "claude-agent-sdk>=0.1.0",

    # GCP SDK
    "google-cloud-run>=0.10.0",
    "google-cloud-storage>=2.14.0",
    "google-cloud-billing>=1.13.0",
    "google-auth>=2.25.0",

    # Data models
    "pydantic>=2.5.0",
    "pyyaml>=6.0",

    # HTTP client
    "httpx>=0.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "pyright>=1.1.0",
]

[project.scripts]
infera = "infera.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/infera"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
